% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anndata2cds}
\alias{anndata2cds}
\title{Convert AnnData object to monocle3 CDS object}
\usage{
anndata2cds(
  inFile,
  outFile = NULL,
  main_layer = "X",
  pcaName = "X_pca",
  umapName = "X_umap"
)
}
\arguments{
\item{inFile}{Path to an input AnnData object on disk (str)}

\item{outFile}{Save output CDS to this file if specified (str or NULL)}

\item{main_layer}{Name of the slot in AnnData to be converted, can be "X",
"raw", default "X" (str)}

\item{pcaName}{Key of PCA in AnnData.obsm to be transferred (str or NULL)}

\item{umapName}{Key of UMAP in AnnData.obsm to be transferred (str or NULL)}
}
\value{
CDS object
}
\description{
This function converts an AnnData object to a monocle3 CDS object
}
