% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anndata2seurat}
\alias{anndata2seurat}
\title{Convert AnnData object to Seurat object(modified)}
\usage{
anndata2seurat(
  inFile,
  outFile = NULL,
  main_layer = "counts_log1p",
  assay = "RNA",
  use_seurat = FALSE,
  lzf = FALSE,
  target_uns_keys = list()
)
}
\arguments{
\item{inFile}{Path to an input AnnData object on disk (str)}

\item{outFile}{Save output Seurat to this file if specified (str or NULL)}

\item{main_layer}{Name of slot in `assay` to store AnnData.X, can be "counts_log1p"
"counts", "data", "scale.data", default "counts_log1p" (str)}

\item{assay}{Name of assay in Seurat object to store expression values,
default "RNA" (str)}

\item{use_seurat}{Use Seurat::ReadH5AD() to do the conversion, default FALSE (logical)}

\item{lzf}{Whether AnnData is compressed by `lzf`, default FALSE (logical)}
}
\value{
Seurat object
}
\description{
This function converts an AnnData object to a Seurat object
}
