% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellratioplot.R
\name{CellRatioPlot}
\alias{CellRatioPlot}
\title{Title}
\usage{
CellRatioPlot(
  object = NULL,
  sample.name = NULL,
  celltype.name = NULL,
  col.width = 0.7,
  flow.alpha = 0.25,
  flow.curve = 0.1,
  fill.col = NULL
)
}
\arguments{
\item{object}{seurat object}

\item{sample.name}{the group (treat vs ctrl)}

\item{celltype.name}{the celltype}

\item{col.width}{the width of col}

\item{flow.alpha}{the alpha of flow}

\item{flow.curve}{the curve of flow}

\item{fill.col}{the color you filled in}
}
\value{
p
}
\description{
Title
}
\examples{
p = CellRatioPlot(object = scRNA,sample.name = "Type",celltype.name = "celltype",fill.col = ov_palette)  ggsave(p,filename=paste0("preanalysis",'/ratioplot.pdf'),width = 5,height = 6)
}
