% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_methods.R
\name{convertFormat}
\alias{convertFormat}
\title{Convert between data objects}
\usage{
convertFormat(
  obj,
  from = c("anndata", "seurat", "sce", "loom"),
  to = c("anndata", "loom", "sce", "seurat", "cds"),
  outFile = NULL,
  main_layer = NULL,
  ...
)
}
\arguments{
\item{obj}{Input a scRNAobject (Maybe a anndata, Seurat object)}

\item{from}{Format of input object, e.g. "anndata", "seurat", "sce", "loom",
etc (str)}

\item{to}{Format of output object (str)}

\item{main_layer}{Required by some formats, may be "counts", "data",
"scale.data", etc (str)}
}
\value{
Output object
}
\description{
This function converts between data format frequently used to hold single
cell data
}
