% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.obs2metadata}
\alias{.obs2metadata}
\title{Prepare cell metadata}
\usage{
.obs2metadata(obs_pd, assay = "RNA")
}
\arguments{
\item{obs_pd}{Input AnnData.obs dataframe}

\item{assay}{Assay name, default "RNA" (str)}
}
\value{
AnnData object
}
\description{
This function prepare cell metadata from AnnData.obs
}
