% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.regularise_df}
\alias{.regularise_df}
\title{Regularise dataframe}
\usage{
.regularise_df(df, drop_single_values = TRUE)
}
\arguments{
\item{df}{Input data frame, usually cell metadata table (data.frame-like
object)}

\item{drop_single_values}{Drop columns with only a single value (logical)}
}
\value{
Dataframe
}
\description{
This function checks if certain columns of a dataframe is of a single value
and drop them if required
}
