% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.var2feature_metadata}
\alias{.var2feature_metadata}
\title{Prepare feature metadata}
\usage{
.var2feature_metadata(var_pd)
}
\arguments{
\item{var_pd}{Input AnnData.var dataframe}
}
\value{
AnnData object
}
\description{
This function prepare feature metadata from AnnData.var
}
