% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loom2sce}
\alias{loom2sce}
\title{Read a Loom file into a SingleCellExperiment object}
\usage{
loom2sce(
  inFile,
  outFile = NULL,
  main_layer = NULL,
  main_layer_name = NULL,
  ...
)
}
\arguments{
\item{inFile}{Path to input Loom file on disk (str)}

\item{outFile}{Save output SingleCellExperiment object to this file if
specified (str or NULL)}
}
\value{
LoomExperiment object
}
\description{
This function reads a Loom file object into a SingleCellExperiment object
}
