% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sce2anndata}
\alias{sce2anndata}
\title{Convert SingleCellExperiment object to AnnData}
\usage{
sce2anndata(
  obj,
  outFile = NULL,
  main_layer = "counts",
  transfer_layers = NULL,
  drop_single_values = TRUE
)
}
\arguments{
\item{obj}{Input SingleCellExperiment object}

\item{outFile}{Save output AnnData to this file if specified (str or NULL)}

\item{main_layer}{Assay to be converted to AnnData.X, default "counts" (str)}

\item{transfer_layers}{If specified, convert additional assays to
AnnData.layers[<slot>], (vector of str)}

\item{drop_single_values}{Drop single value columns in cell metadata table,
default TRUE (logical)}
}
\value{
AnnData object
}
\description{
This function converts a SingleCellExperiment object to an Anndata object
}
