% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sce2loom}
\alias{sce2loom}
\title{Convert SingleCellExperiment object to Loom file}
\usage{
sce2loom(obj, outFile, main_layer = NULL, drop_single_values = TRUE, ...)
}
\arguments{
\item{obj}{Input SingleCellExperiment object}

\item{outFile}{Save output Loom to this file if specified (str or NULL)}

\item{drop_single_values}{Drop single value columns in cell metadata table,
default TRUE (logical)}
}
\value{
LoomExperiment object
}
\description{
This function converts a SingleCellExperiment object to a LoomExperiment
object
}
