% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seurat2anndata}
\alias{seurat2anndata}
\title{Convert Seurat object to AnnData}
\usage{
seurat2anndata(
  obj,
  outFile = NULL,
  assay = "RNA",
  main_layer = "data",
  transfer_layers = NULL,
  drop_single_values = TRUE
)
}
\arguments{
\item{obj}{Input Seurat object}

\item{outFile}{Save output AnnData to this file if specified (str or NULL)}

\item{assay}{Assay to be converted, default "RNA" (str)}

\item{main_layer}{Slot in `assay` to be converted to AnnData.X, may be
"counts", "data", "scale.data", default "data" (str)}

\item{transfer_layers}{If specified, convert slots to AnnData.layers[<slot>],
(vector of str)}

\item{drop_single_values}{Drop single value columns in cell metadata table,
default TRUE (logical)}
}
\value{
AnnData object
}
\description{
This function converts a Seurat object to an Anndata object
}
