% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seurat2sce}
\alias{seurat2sce}
\title{Convert Seurat object to SingleCellExperiment}
\usage{
seurat2sce(obj, outFile = NULL, main_layer = NULL, assay = "RNA", ...)
}
\arguments{
\item{obj}{Input Seurat object}

\item{outFile}{Save output SingleCellExperiment object to this file if
specified (str or NULL)}

\item{assay}{Assay to be converted to AnnData.X, default "RNA" (str)}
}
\value{
AnnData object
}
\description{
This function converts a Seurat object to a SingleCellExperiment object
}
